fix driver

+++ driver/mloUnfoldHost.hpp
@@ -37,10 +37,10 @@ int32_t mloUnFoldFwd4DRunHost(Tgpu* input,
                               const miopenTensorDescriptor_t inputDesc,
                               Tcheck* ref_output,
                               const miopenTensorDescriptor_t ref_outputDesc,
-                              const std::vector<int64_t> kernel_size,
-                              const std::vector<int64_t> stride,
-                              const std::vector<int64_t> padding,
-                              const std::vector<int64_t> dilation)
+                              const std::vector<uint64_t> kernel_size,
+                              const std::vector<uint64_t> stride,
+                              const std::vector<uint64_t> padding,
+                              const std::vector<uint64_t> dilation)
 {
     auto input_tv   = miopen::get_inner_expanded_tv<4>(miopen::deref(inputDesc));
     auto output_tv  = miopen::get_inner_expanded_tv<3>(miopen::deref(ref_outputDesc));
@@ -49,42 +49,42 @@ int32_t mloUnFoldFwd4DRunHost(Tgpu* input,
 
     const int LOCAL_SIZE = 256;
     int spatial_dim_size = input_size - 2;
-    const int64_t N      = static_cast<int64_t>(input_dims[0]);
-    const int64_t C      = static_cast<int64_t>(input_dims[1]);
-    int64_t P = 1, L = 1;
-    std::vector<int64_t> ls;
+    const uint64_t N     = static_cast<uint64_t>(input_dims[0]);
+    const uint64_t C     = static_cast<uint64_t>(input_dims[1]);
+    uint64_t P = 1, L = 1;
+    std::vector<uint64_t> ls;
     for(int i = 0; i < spatial_dim_size; ++i)
     {
         P *= kernel_size[i];
-        int64_t l = (static_cast<int64_t>(input_dims[i + 2]) + 2 * padding[i] -
-                     dilation[i] * (kernel_size[i] - 1) - 1) /
-                        stride[i] +
-                    1;
+        uint64_t l = (static_cast<uint64_t>(input_dims[i + 2]) + 2 * padding[i] -
+                      dilation[i] * (kernel_size[i] - 1) - 1) /
+                         stride[i] +
+                     1;
         L *= l;
         ls.push_back(l);
     }
-    int64_t kernel_size_w = kernel_size[1];
-    int64_t stride_h      = stride[0];
-    int64_t stride_w      = stride[1];
-    int64_t padding_h     = padding[0];
-    int64_t padding_w     = padding[1];
-    int64_t dilation_h    = dilation[0];
-    int64_t dilation_w    = dilation[1];
-    int64_t LW            = ls[1];
-    int64_t H             = static_cast<int64_t>(input_dims[2]);
-    int64_t W             = static_cast<int64_t>(input_dims[3]);
-    int work_size         = (((N * C * P * L) + LOCAL_SIZE - 1) / LOCAL_SIZE) * LOCAL_SIZE;
-    par_ford(work_size)([&](int gid) {
-        int ncp = gid / L, l = gid % L;
-        int nc = ncp / P, p = ncp % P;
-        int n = nc / C, c = nc % C;
+    uint64_t kernel_size_w = kernel_size[1];
+    uint64_t stride_h      = stride[0];
+    uint64_t stride_w      = stride[1];
+    uint64_t padding_h     = padding[0];
+    uint64_t padding_w     = padding[1];
+    uint64_t dilation_h    = dilation[0];
+    uint64_t dilation_w    = dilation[1];
+    uint64_t LW            = ls[1];
+    uint64_t H             = static_cast<uint64_t>(input_dims[2]);
+    uint64_t W             = static_cast<uint64_t>(input_dims[3]);
+    uint64_t work_size     = (((N * C * P * L) + LOCAL_SIZE - 1) / LOCAL_SIZE) * LOCAL_SIZE;
+    par_ford(work_size)([&](uint64_t gid) {
+        uint64_t ncp = gid / L, l = gid % L;
+        uint64_t nc = ncp / P, p = ncp % P;
+        uint64_t n = nc / C, c = nc % C;
         if(n >= N)
             return;
 
-        int lh = l / LW, lw = l % LW;                       // sliding window position
-        int ph = p / kernel_size_w, pw = p % kernel_size_w; // position inside kernel
-        int h = lh * stride_h - padding_h + ph * dilation_h;
-        int w = lw * stride_w - padding_w + pw * dilation_w;
+        int64_t lh = l / LW, lw = l % LW;                       // sliding window position
+        int64_t ph = p / kernel_size_w, pw = p % kernel_size_w; // position inside kernel
+        int64_t h = lh * stride_h - padding_h + ph * dilation_h;
+        int64_t w = lw * stride_w - padding_w + pw * dilation_w;
 
         Tgpu x = static_cast<Tgpu>(0.0f);
         if(0 <= h && h < H && 0 <= w && w < W)
@@ -107,10 +107,10 @@ int32_t mloUnFoldBwd4DRunHost(Tcheck* ref_dinput,
                               const miopenTensorDescriptor_t dinputDesc,
                               Tgpu* doutput,
                               const miopenTensorDescriptor_t doutputDesc,
-                              const std::vector<int64_t> kernel_size,
-                              const std::vector<int64_t> stride,
-                              const std::vector<int64_t> padding,
-                              const std::vector<int64_t> dilation)
+                              const std::vector<uint64_t> kernel_size,
+                              const std::vector<uint64_t> stride,
+                              const std::vector<uint64_t> padding,
+                              const std::vector<uint64_t> dilation)
 {
     auto input_grad_tv   = miopen::get_inner_expanded_tv<4>(miopen::deref(dinputDesc));
     auto output_grad_tv  = miopen::get_inner_expanded_tv<3>(miopen::deref(doutputDesc));
@@ -119,53 +119,53 @@ int32_t mloUnFoldBwd4DRunHost(Tcheck* ref_dinput,
 
     const int LOCAL_SIZE = 256;
     int spatial_dim_size = input_size - 2;
-    const int64_t N      = static_cast<int64_t>(input_grad_dims[0]);
-    const int64_t C      = static_cast<int64_t>(input_grad_dims[1]);
-    int64_t P            = 1;
-    std::vector<int64_t> ls;
+    const uint64_t N     = static_cast<uint64_t>(input_grad_dims[0]);
+    const uint64_t C     = static_cast<uint64_t>(input_grad_dims[1]);
+    uint64_t P           = 1;
+    std::vector<uint64_t> ls;
     for(int i = 0; i < spatial_dim_size; ++i)
     {
         P *= kernel_size[i];
-        int64_t l = (static_cast<int64_t>(input_grad_dims[i + 2]) + 2 * padding[i] -
-                     dilation[i] * (kernel_size[i] - 1) - 1) /
-                        stride[i] +
-                    1;
+        uint64_t l = (static_cast<uint64_t>(input_grad_dims[i + 2]) + 2 * padding[i] -
+                      dilation[i] * (kernel_size[i] - 1) - 1) /
+                         stride[i] +
+                     1;
         ls.push_back(l);
     }
-    int64_t kernel_size_h = kernel_size[0];
-    int64_t kernel_size_w = kernel_size[1];
-    int64_t stride_h      = stride[0];
-    int64_t stride_w      = stride[1];
-    int64_t padding_h     = padding[0];
-    int64_t padding_w     = padding[1];
-    int64_t dilation_h    = dilation[0];
-    int64_t dilation_w    = dilation[1];
-    int64_t LH            = ls[0];
-    int64_t LW            = ls[1];
-    int64_t H             = static_cast<int64_t>(input_grad_dims[2]);
-    int64_t W             = static_cast<int64_t>(input_grad_dims[3]);
-    int work_size         = (((N * C * H * W) + LOCAL_SIZE - 1) / LOCAL_SIZE) * LOCAL_SIZE;
-    par_ford(work_size)([&](int gid) {
-        int nch = gid / W, w = gid % W;
-        int nc = nch / H, h = nch % H;
-        int n = nc / C, c = nc % C;
+    uint64_t kernel_size_h = kernel_size[0];
+    uint64_t kernel_size_w = kernel_size[1];
+    uint64_t stride_h      = stride[0];
+    uint64_t stride_w      = stride[1];
+    uint64_t padding_h     = padding[0];
+    uint64_t padding_w     = padding[1];
+    uint64_t dilation_h    = dilation[0];
+    uint64_t dilation_w    = dilation[1];
+    uint64_t LH            = ls[0];
+    uint64_t LW            = ls[1];
+    uint64_t H             = static_cast<uint64_t>(input_grad_dims[2]);
+    uint64_t W             = static_cast<uint64_t>(input_grad_dims[3]);
+    uint64_t work_size     = (((N * C * H * W) + LOCAL_SIZE - 1) / LOCAL_SIZE) * LOCAL_SIZE;
+    par_ford(work_size)([&](uint64_t gid) {
+        uint64_t nch = gid / W, w = gid % W;
+        uint64_t nc = nch / H, h = nch % H;
+        uint64_t n = nc / C, c = nc % C;
         if(n >= N)
             return;
 
         float sum = 0.0f;
 
-        for(int ph = 0; ph < kernel_size_h; ++ph)
+        for(uint64_t ph = 0; ph < kernel_size_h; ++ph)
         {
-            for(int pw = 0; pw < kernel_size_w; ++pw)
+            for(uint64_t pw = 0; pw < kernel_size_w; ++pw)
             {
-                int lhsh = h - ph * dilation_h + padding_h;
-                int lwsw = w - pw * dilation_w + padding_w;
+                int64_t lhsh = h - ph * dilation_h + padding_h;
+                int64_t lwsw = w - pw * dilation_w + padding_w;
                 if(lhsh % stride_h != 0)
                     continue;
                 if(lwsw % stride_w != 0)
                     continue;
-                int lh = lhsh / stride_h;
-                int lw = lwsw / stride_w;
+                int64_t lh = lhsh / stride_h;
+                int64_t lw = lwsw / stride_w;
                 if(lh < 0 || LH <= lh)
                     continue;
                 if(lw < 0 || LW <= lw)
